requirements.txt
fastapi
uvicorn
reportlab
pandas
numpy
yfinance
ta
beautifulsoup4
selenium
webdriver-manager
python-dotenv
google-generativeai
langchain
langchain-google-genai
requests
aiohttp
python-multipart
pydantic
typing-extensions 
sse-starlette
